04-py-scheduled-jobs/requirements.txt
fastapi
jetpack-io>=0.5.1.dev202201241643053138
uvicorn
python-dotenv
requests